+++ templates/demo-store/app/lib/session.server.ts
@@ -2,7 +2,7 @@ import {
   createCookieSessionStorage,
   type SessionStorage,
   type Session,
-} from '@remix-run/oxygen';
+} from '@shopify/remix-oxygen';
 
 /**
  * This is a custom session implementation for your Hydrogen shop.