+++ templates/demo-store/app/routes/account/__private/orders.$id.tsx
@@ -5,7 +5,7 @@ import {
   json,
   type MetaFunction,
   type LoaderArgs,
-} from '@remix-run/oxygen';
+} from '@shopify/remix-oxygen';
 import {useLoaderData} from '@remix-run/react';
 import {Money, Image, flattenConnection} from '@shopify/hydrogen-react';
 import {statusMessage} from '~/lib/utils';