+++ templates/demo-store/app/routes/cart.tsx
@@ -2,7 +2,11 @@ import {CartLoading, Cart} from '~/components';
 import {Await, useFetcher, useMatches} from '@remix-run/react';
 import {Suspense, useEffect} from 'react';
 import invariant from 'tiny-invariant';
-import {json, type ActionArgs, type AppLoadContext} from '@remix-run/oxygen';
+import {
+  json,
+  type ActionArgs,
+  type AppLoadContext,
+} from '@shopify/remix-oxygen';
 import type {
   Cart as CartType,
   CartInput,
@@ -174,7 +178,8 @@ export async function action({request, context}: ActionArgs) {
     headers.set('Location', redirectTo);
   }
 
-  return json(result, {status, headers});
+  const {cart, errors} = result;
+  return json({cart, errors}, {status, headers});
 }
 
 export default function CartRoute() {