+++ templates/demo-store/remix.config.js
@@ -1,5 +1,5 @@
 // TODO: Change package name when we decide on a package name.
-const {hydrogenRoutes} = require('@shopify/h2-test-hydrogen/build');
+const {hydrogenRoutes} = require('@shopify/hydrogen/build');
 
 /** @type {import('@remix-run/dev').AppConfig} */
 module.exports = {