+++ templates/hello-world/remix.env.d.ts
@@ -1,5 +1,5 @@
 /// <reference types="@remix-run/dev" />
-/// <reference types="@remix-run/oxygen" />
+/// <reference types="@shopify/remix-oxygen" />
 /// <reference types="@shopify/oxygen-workers-types" />
 
 import type {StorefrontClient} from '@shopify/hydrogen';
@@ -20,7 +20,7 @@ declare global {
 /**
  * Declare local additions to `AppLoadContext` to include the session utilities we injected in `server.ts`.
  */
-declare module '@remix-run/oxygen' {
+declare module '@shopify/remix-oxygen' {
   export interface AppLoadContext {
     session: HydrogenSession;
     storefront: StorefrontClient['storefront'];