+++ packages/remix-oxygen/CHANGELOG.md
@@ -1,4 +1,18 @@
+# @shopify/remix-oxygen
+
+## 1.0.0-alpha.1
+
+### Patch Changes
+
+- Initial release of @shopify/remix-oxygen by [@jplhomer](https://github.com/jplhomer)
+
+## 0.0.0-alpha.0
+
+Package renaming. Similar to `@shopify/h2-test-remix-oxygen@0.0.5`.
+
 # @shopify/h2-test-remix-oxygen
+
+> Deprecated package used during early development
 
 ## 0.0.5